moving stuff to utils.py for easy testing
